move aio_ctx to f3
