+++ nvflare/fuel/f3/drivers/aio_grpc_driver.py
@@ -31,7 +31,7 @@ from nvflare.fuel.f3.drivers import net_utils
 from nvflare.fuel.f3.drivers.grpc.streamer_pb2_grpc import (
     StreamerServicer, add_StreamerServicer_to_server, StreamerStub
 )
-from .aio_context import AioContext
+from nvflare.fuel.f3.aio_context import AioContext
 from .base_driver import BaseDriver
 from .grpc.streamer_pb2 import Frame
 from .net_utils import get_address, ssl_required
